+++ trailscraper/s3_download.py
@@ -11,11 +11,23 @@ def _s3_key_prefix(prefix, date, account_id, region):
     return "{}AWSLogs/{}/CloudTrail/{}/{}/{:02d}/{:02d}" \
         .format(prefix, account_id, region, date.year, date.month, date.day)
 
+def _s3_key_prefix_for_org_trails(prefix, date, org_id, account_id, region):
+    return "{}AWSLogs/{}/{}/CloudTrail/{}/{}/{:02d}/{:02d}" \
+        .format(prefix, org_id, account_id, region, date.year, date.month, date.day)
 
-def _s3_key_prefixes(prefix, account_ids, regions, from_date, to_date):
+
+# pylint: disable=too-many-arguments
+def _s3_key_prefixes(prefix, org_ids, account_ids, regions, from_date, to_date):
     delta = to_date.astimezone(pytz.utc) - from_date.astimezone(pytz.utc)
 
     days = [to_date - datetime.timedelta(days=delta_days) for delta_days in range(delta.days + 1)]
+    if org_ids:
+        return [_s3_key_prefix_for_org_trails(prefix, day, org_id, account_id, region)
+                for org_id in org_ids
+                for account_id in account_ids
+                for day in days
+                for region in regions]
+
     return [_s3_key_prefix(prefix, day, account_id, region)
             for account_id in account_ids
             for day in days
@@ -53,8 +65,8 @@ def _s3_download_recursive(bucket, prefix, target_dir):
 
 
 # pylint: disable=too-many-arguments
-def download_cloudtrail_logs(target_dir, bucket, cloudtrail_prefix, account_ids, regions, from_date, to_date):
+def download_cloudtrail_logs(target_dir, bucket, cloudtrail_prefix, org_ids, account_ids, regions, from_date, to_date):
     """Downloads cloudtrail logs matching the given arguments to the target dir"""
-    for prefix in _s3_key_prefixes(cloudtrail_prefix, account_ids, regions, from_date, to_date):
+    for prefix in _s3_key_prefixes(cloudtrail_prefix, org_ids, account_ids, regions, from_date, to_date):
         logging.debug("Downloading logs for %s", prefix)
         _s3_download_recursive(bucket, prefix, target_dir)